# public/openapi.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "HelpDesk AI API",
        "version": "1.0.0",
        "description": "Auto-generated from routes/api.php. Auth: Sanctum Bearer token or scoped API key (X-API-Key). Envelope: {success, data, message}."
    },
    "servers": [
        {
            "url": "http://localhost:8000/api"
        }
    ],
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "Sanctum|API-Key"
            }
        }
    },
    "paths": {
...
        "/api/conversations/{conversation}/message": {
            "post": {
                "operationId": "api.conversations.message",
                "summary": "ConversationController@message",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "tags": [
                    "Conversations"
                ],
                "parameters": [
                    {
                        "name": "conversation",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK \u2014 envelope {success, data, message}"
                    },
                    "201": {
                        "description": "Created"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
...
        "/api/knowledge/categories": {
            "get": {
                "operationId": "api.knowledge.categories",
                "summary": "KnowledgeController@categories",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "tags": [
                    "Knowledge"
                ],
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "OK \u2014 envelope {success, data, message}"
                    },
                    "201": {
                        "description": "Created"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/knowledge/category/{category}": {
            "get": {
                "operationId": "api.knowledge.category",
                "summary": "KnowledgeController@categoryArticles",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "tags": [
                    "Knowledge"
                ],
                "parameters": [
                    {
                        "name": "category",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK \u2014 envelope {success, data, message}"
                    },
                    "201": {
                        "description": "Created"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
...
        "/api/tickets/{ticket}": {
            "get": {
                "operationId": "tickets.show",
                "summary": "TicketController@show",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "tags": [
                    "Tickets"
                ],
                "parameters": [
                    {
                        "name": "ticket",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK \u2014 envelope {success, data, message}"
                    },
                    "201": {
                        "description": "Created"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            },
            "put": {
                "operationId": "tickets.update",
                "summary": "TicketController@update",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "tags": [
                    "Tickets"
                ],
                "parameters": [
                    {
                        "name": "ticket",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK \u2014 envelope {success, data, message}"
                    },
                    "201": {
                        "description": "Created"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            },
            "patch": {
                "operationId": "tickets.update",
                "summary": "TicketController@update",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "tags": [
                    "Tickets"
                ],
                "parameters": [
                    {
                        "name": "ticket",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK \u2014 envelope {success, data, message}"
                    },
                    "201": {
                        "description": "Created"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            },
            "delete": {
                "operationId": "tickets.destroy",
                "summary": "TicketController@destroy",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "tags": [
                    "Tickets"
                ],
                "parameters": [
                    {
                        "name": "ticket",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK \u2014 envelope {success, data, message}"
                    },
                    "201": {
                        "description": "Created"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
...
    }
}